requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pyodbc>=4.0.39
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
schedule>=1.2.0
cryptography>=41.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
